[libgrilio] Added grilio_channel_drop_request. JB#38998
It does basically the same thing as as grilio_channel_cancel_request
but also removes the cancelled request from the pending list so that
it no longer prevents blocking requests from being submitted. This
should only be used for those requests which RIL doesn't bother to
reply to (which is not supposed to happen but unfortunately does).