[libgrilio] Add set_channel callback to GRilIoTransportClass. JB#44551
It's called with non-NULL channel after GRilIoChannel has
been associated with the transport and fully initialized.

The channel is NULL when GRilIoChannel is being deallocated
and therefore disassociated from the transport (right before
the channel drops its reference to the transport which may not
be the last one though).

It also allowed to simplify internal id generation API although
that wasn't the primary reason for adding this callback, it's more
like a nice side effect.